pyproject.toml
[project]
name = "for-test"
version = "0.1.0"
description = "for-test"
readme = "README.md"
requires-python = ">=3.9, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = [
    "plugins"
]
builtin_plugins = ["echo"]

[tool.poetry]
name = "nonebot-starter"
version = "0.1.0"
description = ""
authors = ["Hanjiang <17830831606@163.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nonebot2 = {extras = ["fastapi"], version = "^2.4.0"}
nonebot-adapter-onebot = "^2.4.6"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
